fix(z23+bb08): split FHSS calculator into server page + client component
`app/tools/fhss-calculator/page.tsx` was marked `"use client"` but the
metadata coverage gate (`__tests__/lib/metadata-coverage.test.ts`) requires
every public page to export `metadata` or `generateMetadata()` — which is
only possible from Server Components in Next.js App Router.

Split into:
- `page.tsx` (Server Component) — exports `metadata`, JSON-LD scripts, renders
  `<FHSSCalculatorClient>` via `<Suspense>`
- `FHSSCalculatorClient.tsx` ("use client") — all interactive state, range
  inputs, result cards, compliance footer

Same pattern as `app/tools/withholding-tax-calculator/` (server page +
WithholdingTaxClient).

Fixes: metadata-coverage test failure that caused CI to fail on PR #895.

Refs: #221
Queue: Z-23 + BB-08 CI-rescue iter 440
